Move the (bobp) check to the end of the (cursorcolumn-show) loop
Bug fix: Previously, the (bobp) check in the (cursorcolumn-show)
function was positioned incorrectly, causing a bug where the vertical
line was not appearing in the first line. By moving the (bobp) check to
the end of the function, this issue is now resolved, and the vertical
line is correctly displayed even in the first line.